backend/requirements.txt
anyio==3.6.2
bcrypt==3.2.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.95.1
greenlet==3.1.0
h11==0.14.0
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
passlib==1.7.4
Pillow==9.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.7
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.34
starlette==0.26.1
typing_extensions==4.12.2
urllib3==1.26.15
uvicorn==0.21.1
